added integer index
